+++ src/main/kotlin/com/github/minasploit/migrationfacilitator/Constants.kt
@@ -4,3 +4,5 @@ const val PREFIX = "com.github.minasploit.migrationfacilitator."
 const val STARTUP_PROJECT = PREFIX + "startup_project"
 const val DATA_PROJECT = PREFIX + "data_project"
 const val NOTIFICATION_GROUP_ID = "Migration Facilitator"
+
+const val REMOVE_ALL_MIGRATIONS = "REMOVE ALL APPLIED MIGRATIONS"