+++ src/main/kotlin/com/github/minasploit/migrationfacilitator/Util.kt
@@ -1,8 +1,13 @@
 package com.github.minasploit.migrationfacilitator
 
+import com.github.minasploit.migrationfacilitator.actions.AddMigration
+import com.github.minasploit.migrationfacilitator.actions.RemoveMigration
+import com.github.minasploit.migrationfacilitator.actions.UpdateDatabase
 import com.intellij.notification.NotificationType
 import com.intellij.notification.Notification
+import com.intellij.openapi.fileEditor.impl.LoadTextUtil
 import com.intellij.openapi.project.Project
+import com.intellij.openapi.vfs.LocalFileSystem
 import com.intellij.openapi.vfs.VirtualFileManager
 import org.apache.commons.lang.StringUtils
 import java.io.File
@@ -25,10 +30,17 @@ class Util private constructor() {
             project: Project,
             command: String,
             skipLines: Int = 2,
-            dir: String = project.basePath!!
+            dir: String = project.basePath!!,
+            shouldRefreshFileManager: Boolean = true
         ): Triple<Boolean, String, String> {
-            val process = Runtime.getRuntime().exec(command, null, File(dir))
-            val successful = process.waitFor() == 0
+            val process: Process
+            val successful: Boolean
+            try {
+                process = Runtime.getRuntime().exec(command, null, File(dir))
+                successful = process.waitFor() == 0
+            } catch (ex: Exception) {
+                return Triple(false, "", "")
+            }
 
             val scanner = Scanner(process.inputStream)
 
@@ -56,7 +68,7 @@ class Util private constructor() {
                     val errorProject = StringUtils.substringBetween(output, "Switch: ", "\n")
 
                     errorMessage =
-                        "The project '$errorProject' doesn't exist or is not an SDK-style project. Please make sure it's a valid project type."
+                        "The project '$errorProject' doesn't exist. Please make sure it's a valid project type."
                 }
                 output.contains("Unable to create an object of type") -> {
                     errorMessage =
@@ -69,19 +81,70 @@ class Util private constructor() {
                     errorMessage =
                         "Could not load '$dataProject' as a data project. Ensure it is referenced by the startup project '$startupProject'"
                 }
+                output.contains("doesn't match your migrations assembly") -> {
+                    val dataProject = StringUtils.substringBetween(output, "assembly '", "'")
+                    val userDataProject = StringUtils.substringBetween(output, "project '", "'")
+
+                    errorMessage =
+                        "The specified data project '$userDataProject' doesn't match the set up migration project '$dataProject'. Please use '$dataProject' as the data project."
+                }
+                output.contains("An error occurred while accessing the database.") -> {
+                    errorMessage =
+                        "An error occurred while accessing the database. Make sure you can access the database to ensure correct results."
+                }
+                output.toLowerCase().contains("no migrations were found") -> {
+                    errorMessage = "No Migrations found."
+                }
                 else -> {
-                    errorMessage = ""
+                    errorMessage = "Unknown error occurred. Try selecting the correct startup and data projects."
                 }
             }
 
-            if (successful)
+            if (successful && shouldRefreshFileManager)
                 VirtualFileManager.getInstance().refreshWithoutFileWatcher(true)
 
-            return Triple(successful, outputTrimmed, errorMessage)
+            return Triple(successful, outputTrimmed.trim(), errorMessage.trim())
         }
 
-        fun buildDotnetCommand(commandSection: String, startupProject: String, dataProject: String): String {
-            return "dotnet ef $commandSection -s \"$startupProject\" -p \"$dataProject\""
+        fun buildDotnetCommand(
+            commandSection: String,
+            startupProject: String,
+            dataProject: String,
+            build: Boolean = true
+        ): String {
+            var command = "dotnet ef $commandSection -s \"$startupProject\" -p \"$dataProject\""
+            if (!build)
+                command += " -- no-build"
+
+            return command
+        }
+
+        fun enableAllButtons() {
+            changeButtonsStatus(true)
+        }
+
+        fun disableAllButtons() {
+            changeButtonsStatus(false)
+        }
+
+        private fun changeButtonsStatus(enable: Boolean) {
+            AddMigration.IsEnabled = enable
+            RemoveMigration.IsEnabled = enable
+            UpdateDatabase.IsEnabled = enable
+        }
+
+        fun getProjectsInSolution(project: Project): List<String> {
+            return try {
+                val solutionDirectoryVirtualFile = LocalFileSystem.getInstance().findFileByPath(project.basePath!!)!!
+                val solutionVirtualFile = solutionDirectoryVirtualFile.children.first { it.name.contains(".sln") }
+                val solutionContent = LoadTextUtil.loadText(solutionVirtualFile).toString()
+                val projectsDetail = StringUtils.substringsBetween(solutionContent, "Project", "EndProject")
+                projectsDetail.map {
+                    StringUtils.substringBetween(it, "= \"", "\"")
+                }
+            } catch (ex: Exception) {
+                arrayListOf(project.name)
+            }
         }
     }
 }