+++ src/main/kotlin/com/github/minasploit/migrationfacilitator/actions/UpdateDatabaseDialog.kt
@@ -1,15 +1,16 @@
 package com.github.minasploit.migrationfacilitator.actions
 
+import com.github.minasploit.migrationfacilitator.BaseDialogWrapper
+import com.github.minasploit.migrationfacilitator.REMOVE_ALL_MIGRATIONS
 import com.github.minasploit.migrationfacilitator.DATA_PROJECT
 import com.github.minasploit.migrationfacilitator.STARTUP_PROJECT
 import com.github.minasploit.migrationfacilitator.Util
-import com.intellij.ide.util.PropertiesComponent
 import com.intellij.notification.NotificationType
 import com.intellij.openapi.progress.ProgressIndicator
 import com.intellij.openapi.progress.ProgressManager
 import com.intellij.openapi.progress.Task
 import com.intellij.openapi.project.Project
-import com.intellij.openapi.ui.DialogWrapper
+import com.intellij.openapi.ui.Messages
 import java.awt.Dimension
 import java.awt.event.ActionEvent
 import javax.swing.AbstractAction
@@ -18,15 +19,9 @@ import javax.swing.JButton
 import javax.swing.JComponent
 import javax.swing.JLabel
 import javax.swing.JPanel
-import javax.swing.JSeparator
-import javax.swing.JTextField
-import javax.swing.SwingUtilities
 
-class UpdateDatabaseDialog(private val project: Project) : DialogWrapper(project, true) {
-    private val properties: PropertiesComponent = PropertiesComponent.getInstance(project)
+class UpdateDatabaseDialog(private val project: Project) : BaseDialogWrapper(project, true) {
     private val migrationSelector = com.intellij.openapi.ui.ComboBox<String>()
-    private val startupProjectInput = JTextField(properties.getValue(STARTUP_PROJECT, project.name))
-    private val dataProjectInput = JTextField(properties.getValue(DATA_PROJECT, project.name))
     private val refreshMigrationButton = JButton()
 
     override fun createCenterPanel(): JComponent {
@@ -35,23 +30,6 @@ class UpdateDatabaseDialog(private val project: Project) : DialogWrapper(project
         dialogPanel.preferredSize = Dimension(75, 0)
 
         val description = JLabel("This allows you to update the database to a specific migration")
-//        description.preferredSize = Dimension()
-
-        val startupProjectInputLabel = JLabel("Startup Project. Ex: Solution.StartupProject", JLabel.TRAILING)
-//        startupProjectInputLabel.preferredSize = Dimension()
-
-        val dataProjectInputLabel =
-            JLabel("Data Project, the project where you store the DbContext. Ex: Solution.DataProject", JLabel.TRAILING)
-//        dataProjectInputLabel.preferredSize = Dimension()
-
-        val separator = JSeparator()
-        separator.size = Dimension(10, 20)
-
-        startupProjectInput.preferredSize = Dimension(75, 25)
-        startupProjectInputLabel.labelFor = startupProjectInput
-
-        dataProjectInput.preferredSize = Dimension(75, 25)
-        dataProjectInputLabel.labelFor = dataProjectInput
 
         refreshMigrationButton.action = object : AbstractAction() {
             override fun actionPerformed(e: ActionEvent?) {
@@ -61,17 +39,12 @@ class UpdateDatabaseDialog(private val project: Project) : DialogWrapper(project
         refreshMigrationButton.text = "Refresh Migrations"
         refreshMigrationButton.isVisible = false
 
-        SwingUtilities.invokeLater {
-            refreshMigrationList()
-        }
+        refreshMigrationList()
 
         dialogPanel.add(description)
         dialogPanel.add(separator)
         dialogPanel.add(migrationSelector)
-        dialogPanel.add(startupProjectInputLabel)
-        dialogPanel.add(startupProjectInput)
-        dialogPanel.add(dataProjectInputLabel)
-        dialogPanel.add(dataProjectInput)
+        addDefaultUi(dialogPanel)
         dialogPanel.add(refreshMigrationButton)
 
         return dialogPanel
@@ -79,22 +52,41 @@ class UpdateDatabaseDialog(private val project: Project) : DialogWrapper(project
 
     // fetch all available migrations and add them to the selector
     fun refreshMigrationList() {
-        ProgressManager.getInstance().run(object : Task.Backgroundable(project, "Loading existing migrations...") {
-            override fun run(indicator: ProgressIndicator) {
-                // start your process
+        Util.disableAllButtons()
 
+        ProgressManager.getInstance().runProcessWithProgressSynchronously(
+            {
                 val (success, output, errorMessage) = Util.runCommand(
                     project,
-                    Util.buildDotnetCommand("migrations list", startupProjectInput.text, dataProjectInput.text)
+                    Util.buildDotnetCommand(
+                        "migrations list",
+                        startupProjectSelector.item,
+                        dataProjectSelector.item,
+                        false
+                    ),
+                    2, project.basePath!!, false
                 )
 
                 if (success) {
-                    val migrationItems = mutableListOf<String>()
-                    migrationItems.add("REMOVE ALL APPLIED MIGRATIONS")
+                    var migrationItems = mutableListOf<String>()
+                    migrationItems.add(REMOVE_ALL_MIGRATIONS)
                     migrationItems.addAll(
                         output.split("\n")
-                            .dropLast(1)
                     )
+
+                    if (errorMessage.contains("An error occurred while accessing the database")) {
+                        // remove the error message from the migration list
+
+                        // 0 -> Remove All Migrations
+                        // 1 -> An error occurred while accessing the database...
+                        // 2... -> Other migrations
+
+                        migrationItems.removeAt(1)
+                    } else if (errorMessage == "No Migrations found.") {
+                        // no migrations in the project
+
+                        migrationItems = migrationItems.dropLast(1).toMutableList()
+                    }
 
                     migrationItems.forEach {
                         migrationSelector.addItem(it)
@@ -102,8 +94,8 @@ class UpdateDatabaseDialog(private val project: Project) : DialogWrapper(project
 
                     migrationSelector.selectedIndex = migrationSelector.itemCount - 1
 
-                    properties.setValue(STARTUP_PROJECT, startupProjectInput.text)
-                    properties.setValue(DATA_PROJECT, dataProjectInput.text)
+                    properties.setValue(STARTUP_PROJECT, startupProjectSelector.item)
+                    properties.setValue(DATA_PROJECT, dataProjectSelector.item)
 
                     refreshMigrationButton.isVisible = false
                 } else {
@@ -116,28 +108,47 @@ class UpdateDatabaseDialog(private val project: Project) : DialogWrapper(project
 
                     refreshMigrationButton.isVisible = true
                 }
-            }
-        })
+
+                Util.enableAllButtons()
+            },
+            "Loading existing migrations...", false, project, contentPanel
+        )
     }
 
     override fun doOKAction() {
         val item = migrationSelector.selectedItem as String
 
         val migrationName =
-            if (item == "REMOVE ALL APPLIED MIGRATIONS") "0" else item.split(" ")[0]
+            if (item == REMOVE_ALL_MIGRATIONS) "0" else item.split(" ")[0]
 
-        properties.setValue(STARTUP_PROJECT, startupProjectInput.text)
-        properties.setValue(DATA_PROJECT, dataProjectInput.text)
+        if (migrationName == "0") {
+            val stopOperation = Messages.showOkCancelDialog(
+                contentPanel,
+                "Are you sure you want to reset the database?",
+                "Confirmation",
+                "Continue",
+                "Cancel",
+                Messages.getQuestionIcon()
+            ) != Messages.YES
+
+            if (stopOperation)
+                return
+        }
+
+        Util.disableAllButtons()
+
+        properties.setValue(STARTUP_PROJECT, startupProjectSelector.item)
+        properties.setValue(DATA_PROJECT, dataProjectSelector.item)
 
         ProgressManager.getInstance().run(object : Task.Backgroundable(project, "Updating database...") {
             override fun run(indicator: ProgressIndicator) {
                 // start your process
-                val (success, output) = Util.runCommand(
+                val (success, output, errorMessage) = Util.runCommand(
                     project,
                     Util.buildDotnetCommand(
                         "database update \"$migrationName\"",
-                        startupProjectInput.text,
-                        dataProjectInput.text
+                        startupProjectSelector.item,
+                        dataProjectSelector.item
                     ),
                     0
                 )
@@ -145,18 +156,20 @@ class UpdateDatabaseDialog(private val project: Project) : DialogWrapper(project
                     Util.showNotification(
                         project,
                         "Database updated",
-                        if (migrationName == "0") "All migrations have been removed and the database has been reset."
-                        else "Migration: '$migrationName' applied to database referenced in the project ${dataProjectInput.text}",
+                        if (migrationName == "0") "All migrations have been removed from the database."
+                        else "Migration: '$migrationName' applied to database referenced in the project '${dataProjectSelector.item}'",
                         NotificationType.INFORMATION
                     )
                 } else {
                     Util.showNotification(
                         project,
                         "Can't update database",
-                        output,
+                        if (errorMessage != "") errorMessage else output,
                         NotificationType.ERROR
                     )
                 }
+
+                Util.enableAllButtons()
             }
         })