+++ src/main/kotlin/com/github/minasploit/migrationfacilitator/actions/VersionAction.kt
@@ -1,10 +1,16 @@
 package com.github.minasploit.migrationfacilitator.actions
 
-import com.intellij.openapi.actionSystem.AnAction
+import com.github.minasploit.migrationfacilitator.BaseAction
 import com.intellij.openapi.actionSystem.AnActionEvent
 
-class VersionAction : AnAction() {
+class VersionAction : BaseAction() {
     override fun actionPerformed(e: AnActionEvent) {
         // do nothing
+    }
+
+    override fun update(e: AnActionEvent) {
+        e.presentation.isEnabled = false
+
+        super.update(e)
     }
 }